added features like jwt and protected routing

+++ Backend/server.js
@@ -3,6 +3,7 @@ const express = require('express');
 const Campground = require('./Models/Campground')
 const User = require('./Models/User');
 const dbConnect = require('./Seeds/index');
+const rateLimit = require('express-rate-limit');
 const bcrypt = require('bcryptjs');
 const { v4: uuidv4 } = require('uuid');
 const jwt = require('jsonwebtoken');
@@ -20,18 +21,50 @@ const server = app.listen(PORT, () => {
 });
 const io = socketIo(server, {
   cors: {
-    origin: '*', // Allow any origin
-    methods: ['GET', 'POST']
+    origin: process.env.FRONTEND_SERVER, // Allow any origin
+    methods: ['GET', 'POST'],
+    credentials: true,
   }
 });
-
+// MIDDLEWARES 
 // parse application/x-www-form-urlencoded
 app.use(express.urlencoded({ extended: true }));
 
-// parse application/json
+// parse application/json       
 app.use(express.json());
 app.use(cookieParser());
 app.use(cors({ origin: process.env.FRONTEND_SERVER, credentials: true }));
+const apiLimiter = rateLimit({
+  windowMs: 15 * 60 * 1000, // 15 minutes
+  max: 100, // Limit each IP to 100 requests per windowMs
+});
+app.use('/api/', apiLimiter);
+
+// Helper function to generate JWT tokens
+const generateToken = (user) => {
+  return jwt.sign({ user }, process.env.JWT_SECRET, {
+    expiresIn: '1h', // Tokens expire in 1 hour
+  });
+};
+
+// Authentication middleware
+const authenticateToken = (req, res, next) => {
+  const token = req.cookies.userIdYelp || req.headers['authorization'];
+  if (!token) return res.status(401).json({ message: 'Access denied. No token provided.' });
+  try {
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    req.user = decoded.user
+    req.tokenExpiry = decoded.exp ? decoded.exp : null; 
+    next();
+  } catch (ex) {
+    res.status(400).json({ message: 'Invalid token.', tokenExpiry: req.tokenExpiry });
+  }
+};
+
+app.get('/api/protected', authenticateToken, (req, res) => {
+  // Access req.user and req.tokenExpiry if needed
+  res.json({ message: 'Access granted to protected data.', user: req.user, tokenExpiry: req.tokenExpiry });
+});
 
 app.post('/api/register', async (req, res) => {
   const { username, password } = req.body;
@@ -46,7 +79,8 @@ app.post('/api/register', async (req, res) => {
   });
 
   await newUser.save();
-  res.json({ success: true, userId: newUser.userId });
+  const token = generateToken(username);
+  res.json({ success: true, jwtToken: token });
 });
 
 app.post('/api/login', async (req, res) => {
@@ -57,11 +91,11 @@ app.post('/api/login', async (req, res) => {
 
   const isMatch = await bcrypt.compare(password, user.password);
   if (!isMatch) return res.json({ success: false, message: 'Invalid credentials' });
-
-  res.json({ success: true, userId: user.userId });
+  const token = generateToken(username);
+  res.json({ success: true, jwtToken: token });
 });
 
-app.get('/api/users', async (req, res) => {
+app.get('/api/users', authenticateToken, async (req, res) => {
   try {
     const users = await User.find(); // Fetch all users from the database
     res.json(users);
@@ -70,10 +104,11 @@ app.get('/api/users', async (req, res) => {
   }
 });
 
-app.get('/api/allusers/:userId', async (req, res) => {
+app.get('/api/allusers/:userId', authenticateToken, async (req, res) => {
   try {
-    const username = req.params.userId
-    const user = await User.findOne({username}); // Fetch user by ID
+    const userId = req.params.userId
+    const user = await User.findOne({username: userId}); // Fetch user by ID
+    console.log(user)
     res.json(user);
   } catch (error) {
     console.log(error)
@@ -82,9 +117,9 @@ app.get('/api/allusers/:userId', async (req, res) => {
 });
 
 // Get posts of a specific user
-app.get('/api/users/:userId/posts', async (req, res) => {
+app.get('/api/users/:userId/posts', authenticateToken, async (req, res) => {
   try {
-    const {posts} = await User.findOne({ username: req.params.userId }); // Fetch posts by user ID
+    const {posts} = await User.findOne({ username: req.user }); // Fetch posts by user ID
     res.json(posts);
   } catch (error) {
     res.status(500).send('Server Error');
@@ -93,10 +128,10 @@ app.get('/api/users/:userId/posts', async (req, res) => {
 
 
 //fetching user by userIdYelp
-app.get('/api/users/:userId', async (req, res) => {
+app.get('/api/users/:userId', authenticateToken, async (req, res) => {
   try {
-    const { userId } = req.params; // Get userId from the URL parameters
-    const user = await User.findOne({ userId });
+    const userId  = req.user; // Get userId from the URL parameters
+    const user = await User.findOne({ username: userId });
     if (!user) {
         return res.status(404).json({ message: 'User not found' });
     }
@@ -106,12 +141,12 @@ app.get('/api/users/:userId', async (req, res) => {
 }
 });
 
-app.put('/api/users/:userId', async (req, res) => {
+app.put('/api/users/:userId', authenticateToken, async (req, res) => {
   try {
-    const { userId } = req.params; // Get userId from the URL parameters
+    const userId = req.user; // Get userId from the URL parameters
     const updatedData = req.body; // Get the data to update from the request body
     // Update the user in the database
-    const user = await User.findOneAndUpdate({ userId }, updatedData, {
+    const user = await User.findOneAndUpdate({ username: userId }, updatedData, {
         new: true, // Return the updated document
         runValidators: true, // Ensure validation on update
     });
@@ -127,14 +162,14 @@ app.put('/api/users/:userId', async (req, res) => {
 
 
 // Route to delete a specific post
-app.delete('/api/users/:userId/deletePost', async (req, res) => {
-  const { postId } = req.body; // Extract postId from request body
-  const { userId } = req.params; // Extract userId from request parameters
+app.delete('/api/users/deletePost/:postId', authenticateToken, async (req, res) => {
+  const { postId } = req.params; // Extract postId from request body
+  const userId = req.user; // Extract userId from request parameters
 
   try {
     // Find the user and remove the post with the matching postId
     const postsnum = await User.findOne(
-      { userId }  
+      { username: userId }  
     );
 
     if(postsnum.posts.length === 1){
@@ -142,15 +177,13 @@ app.delete('/api/users/:userId/deletePost', async (req, res) => {
     }
 
         const user = await User.findOneAndUpdate(
-        { userId }, 
+        { username: userId }, 
         { $pull: { posts: { postId } } }, // Remove post with matching postId
         { new: true } // Return the updated user document
       );
-
      await Post.findOneAndDelete(
-      { postId } 
+      { postId: postId } 
     );
-
     res.status(200).json({ message: 'Post deleted successfully' });
   } catch (error) {
     console.log(error)
@@ -159,10 +192,10 @@ app.delete('/api/users/:userId/deletePost', async (req, res) => {
 });
 
 
-app.put('/api/users/:userId/clearBio', async (req, res) => {
+app.put('/api/users/clearBio/:Id', authenticateToken, async (req, res) => {
   try {
-    const { userId } = req.params;
-    const user = await User.findOne({userId});
+    const userId = req.user;
+    const user = await User.findOne({username: userId});
     user.about = ''; // Clear the about field
     await user.save();
     res.status(200).json({ message: 'Bio cleared successfully' });
@@ -173,7 +206,7 @@ app.put('/api/users/:userId/clearBio', async (req, res) => {
 
 
 // Get all posts
-app.get('/api/posts', async (req, res) => {
+app.get('/api/posts', authenticateToken, async (req, res) => {
   try {
     const posts = await Post.find(); // Populating the user’s name
     res.json(posts);
@@ -182,7 +215,7 @@ app.get('/api/posts', async (req, res) => {
   }
 });
 
-app.put('/api/posts/:id/like', async (req, res) => {
+app.put('/api/posts/:id/like', authenticateToken, async (req, res) => {
   try {
     const post = await Post.findById(req.params.id);
     if (!post) return res.status(404).send('Post not found');
@@ -195,13 +228,22 @@ app.put('/api/posts/:id/like', async (req, res) => {
   }
 });
 
-app.get('/api/posts/:id/comments', async (req, res) => {
+app.get('/api/posts/:id/comments', authenticateToken, async (req, res) => {
   try {
       const post = await Post.findById(req.params.id).populate('comments');
       res.status(200).json(post.comments);
   } catch (error) {
       res.status(500).json({ error: 'Error fetching comments' });
   }
+});
+
+
+io.use((socket, next) => {
+  const token = socket.handshake.auth.token;
+  if (token && jwt.verify(token, process.env.JWT_SECRET)) {
+    return next();
+  }
+  return next(new Error('Authentication error'));
 });
 
 // Socket.io connection for real-time comments
@@ -229,10 +271,10 @@ io.on('connection', (socket) => {
 
 
 
-app.put('/api/posts/:userId', async (req, res) => {
+app.put('/api/posts/:userId', authenticateToken, async (req, res) => {
   try {
-    const {userId} = req.params; // Correctly extracting userId
-    const user = await User.findOne({userId}); // Find user by ID
+    const userId = req.user; // Correctly extracting userId
+    const user = await User.findOne({username: userId}); // Find user by ID
     if (!user) {
       return res.status(404).json({ message: 'User not found' });
     }
@@ -272,7 +314,7 @@ app.put('/api/posts/:userId', async (req, res) => {
 });
 
 
-app.get('/campground', async (req, res) => {
+app.get('/campground', authenticateToken, async (req, res) => {
   try {
     const campgrounds = await Campground.find({});
     res.json([...campgrounds]);
@@ -281,20 +323,61 @@ app.get('/campground', async (req, res) => {
   }
 });
 
-app.put('/campground', async(req, res)=>{
-  let id = req.body._id
-  await Campground.findByIdAndUpdate(id,{location:req.body.location, title:req.body.title, price:req.body.price, description:req.body.description})
-  res.send('it worked')
-})
 
-app.post('/campground',async (req, res) => {
+app.put('/campground', authenticateToken, async (req, res) => {
+  const { _id, location, title, price, description } = req.body;
+
+  try {
+    // Find the campground by ID
+    const campground = await Campground.findById(_id);
+
+    // Check if the campground exists
+    if (!campground) {
+      return res.status(404).json({ message: 'Campground not found' });
+    }
+
+    // Check if the user has permission to edit
+    if (campground.username === req.user) {
+      await Campground.findByIdAndUpdate(_id, { location, title, price, description });
+      return res.status(200).json({ message: 'Campground updated successfully' });
+    } else {
+      return res.status(403).json({ message: 'You do not have permission to edit this campground' });
+    }
+  } catch (error) {
+    console.error("Error updating campground:", error);
+    res.status(500).json({ message: 'An error occurred while trying to update the campground' });
+  }
+});
+
+
+app.post('/campground', authenticateToken,async (req, res) => {
   const campground = new Campground( req.body)
+  campground.username = req.user
   await campground.save()
   res.send('it worked')
 })
 
-app.delete('/campground/:id', async(req, res)=>{
+app.delete('/campground/:id', authenticateToken, async (req, res) => {
   const { id } = req.params;
-  await Campground.findByIdAndDelete(id)
-  res.send('it worked')
-})
+
+  try {
+    // Find the campground by ID
+    const campground = await Campground.findById(id);
+    
+    // Check if the campground exists
+    if (!campground) {
+      return res.status(404).json({ message: 'Campground not found' });
+    }
+
+    // Check if the user has permission to delete
+    if (campground.username === req.user) {
+      await Campground.findByIdAndDelete(id);
+      return res.status(200).json({ message: 'Campground deleted successfully' });
+    } else {
+      return res.status(403).json({ message: 'You do not have permission to delete this campground' });
+    }
+  } catch (error) {
+    console.error("Error deleting campground:", error);
+    res.status(500).json({ message: 'An error occurred while trying to delete the campground' });
+  }
+});
